+++ src/main/java/mate/jdbc/service/impl/ManufacturerServiceImpl.java
@@ -3,6 +3,7 @@ package mate.jdbc.service.impl;
 import java.util.List;
 import java.util.NoSuchElementException;
 import mate.jdbc.dao.ManufacturerDao;
+import mate.jdbc.exception.DataProcessingException;
 import mate.jdbc.lib.Inject;
 import mate.jdbc.lib.Service;
 import mate.jdbc.model.Manufacturer;
@@ -15,11 +16,17 @@ public class ManufacturerServiceImpl implements ManufacturerService {
 
     @Override
     public Manufacturer create(Manufacturer manufacturer) {
+        if (manufacturer == null) {
+            throw new DataProcessingException("Can't create manufacturer from null");
+        }
         return manufacturerDao.create(manufacturer);
     }
 
     @Override
     public Manufacturer get(Long id) {
+        if (id == null) {
+            throw new DataProcessingException("Can't get manufacturer from null");
+        }
         return manufacturerDao.get(id)
                 .orElseThrow(() -> new NoSuchElementException("Can't get manufacturer by id: "
                         + id));
@@ -32,11 +39,17 @@ public class ManufacturerServiceImpl implements ManufacturerService {
 
     @Override
     public Manufacturer update(Manufacturer manufacturer) {
+        if (manufacturer == null) {
+            throw new DataProcessingException("Can't update manufacturer from null");
+        }
         return manufacturerDao.update(manufacturer);
     }
 
     @Override
     public boolean delete(Long id) {
+        if (id == null) {
+            throw new DataProcessingException("Can't delete manufacturer null");
+        }
         return manufacturerDao.delete(id);
     }
 }
